+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v3/Information.java
@@ -1,4 +1,4 @@
-package edu.illinois.library.cantaloupe.resource.iiif.v2;
+package edu.illinois.library.cantaloupe.resource.iiif.v3;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
@@ -7,18 +7,16 @@ import java.util.LinkedHashMap;
 import java.util.List;
 
 /**
- * <p>Class whose instances are intended to be serialized to JSON for use in
- * IIIF Image Information responses.</p>
+ * <p>Class whose instances are intended to be serialized as JSON for use in
+ * information responses.</p>
  *
- * <p>Extends Map in order to support arbitrary keys, and LinkedHashMap in
- * order to preserve key order.</p>
+ * <p>Extends {@link LinkedHashMap} in order to support arbitrary keys and
+ * preserve key order.</p>
  *
- * @see <a href="http://iiif.io/api/image/2.0/#image-information">IIIF Image
- * API 2.0: Image Information</a>
- * @see <a href="https://github.com/FasterXML/jackson-databind">jackson-databind
- * docs</a>
+ * @see <a href="https://iiif.io/api/image/3.0/#52-technical-properties">IIIF
+ * Image API 3.0: Technical Properties</a>
  */
-class ImageInfo<K,V> extends LinkedHashMap<K,V> {
+class Information<K,V> extends LinkedHashMap<K,V> {
 
     @JsonPropertyOrder({ "width", "height" })
     public static final class Size {
@@ -29,7 +27,7 @@ class ImageInfo<K,V> extends LinkedHashMap<K,V> {
         public Size() {}
 
         public Size(Integer width, Integer height) {
-            this.width = width;
+            this.width  = width;
             this.height = height;
         }
     }
