+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v3/InformationResource.java
@@ -41,7 +41,7 @@ public class InformationResource extends IIIF3Resource {
     }
 
     /**
-     * Writes a JSON-serialized {@link ImageInfo} instance to the response.
+     * Writes a JSON-serialized {@link Information} instance to the response.
      */
     @Override
     public void doGET() throws Exception {
@@ -121,16 +121,16 @@ public class InformationResource extends IIIF3Resource {
 
     private JacksonRepresentation newRepresentation(Info info,
                                                     Set<Format> availableOutputFormats) {
-        final ImageInfoFactory factory = new ImageInfoFactory();
+        final InformationFactory factory = new InformationFactory();
         factory.setDelegateProxy(getDelegateProxy());
 
-        final ImageInfo<String, Object> imageInfo = factory.newImageInfo(
+        final Information<String, Object> iiifInfo = factory.newImageInfo(
                 availableOutputFormats,
                 getImageURI(),
                 info,
                 getPageIndex(),
                 getMetaIdentifier().getScaleConstraint());
-        return new JacksonRepresentation(imageInfo);
+        return new JacksonRepresentation(iiifInfo);
     }
 
 }
