+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v2/InformationResourceTest.java
@@ -440,7 +440,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://localhost:" + HTTP_PORT +
                 Route.IIIF_2_PATH + "/" + IMAGE, info.get("@id"));
     }
@@ -455,7 +455,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://example.org" +
                 Route.IIIF_2_PATH + "/" + IMAGE, info.get("@id"));
     }
@@ -471,7 +471,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://localhost:" + HTTP_PORT +
                 Route.IIIF_2_PATH + path, info.get("@id"));
     }
@@ -487,7 +487,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://localhost:" + HTTP_PORT +
                 Route.IIIF_2_PATH + path, info.get("@id"));
     }
@@ -505,7 +505,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://example.org:8080/cats" +
                 Route.IIIF_2_PATH + "/originalID", info.get("@id"));
     }
@@ -524,7 +524,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("https://example.net" +
                 Route.IIIF_2_PATH + "/" + IMAGE, info.get("@id"));
     }
