+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v3/InformationResourceTest.java
@@ -415,7 +415,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://localhost:" + HTTP_PORT +
                 Route.IIIF_3_PATH + "/" + IMAGE, info.get("id"));
     }
@@ -430,7 +430,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://example.org" +
                 Route.IIIF_3_PATH + "/" + IMAGE, info.get("id"));
     }
@@ -446,7 +446,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://localhost:" + HTTP_PORT +
                 Route.IIIF_3_PATH + path, info.get("id"));
     }
@@ -462,7 +462,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://localhost:" + HTTP_PORT +
                 Route.IIIF_3_PATH + path, info.get("id"));
     }
@@ -480,7 +480,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("http://example.org:8080/cats" +
                 Route.IIIF_3_PATH + "/originalID", info.get("id"));
     }
@@ -499,7 +499,7 @@ public class InformationResourceTest extends ResourceTest {
 
         String json = response.getBodyAsString();
         ObjectMapper mapper = new ObjectMapper();
-        ImageInfo<?, ?> info = mapper.readValue(json, ImageInfo.class);
+        Information<?, ?> info = mapper.readValue(json, Information.class);
         assertEquals("https://example.net" +
                 Route.IIIF_3_PATH + "/" + IMAGE, info.get("id"));
     }
